# Improve `#members` performance