--- a/Gemfile
+++ b/Gemfile
@@ -9,3 +9,4 @@ gem "rake", "~> 13.0"
 
 gem "minitest", "~> 5.0"
 gem "debug"
+gem "benchmark-ips"