Improve `#members` performance by ~4× at a small cost to `.define`.

Before
```
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [arm64-darwin21]
Warming up --------------------------------------
             .define    20.855k i/100ms
            #members   462.221k i/100ms
Calculating -------------------------------------
             .define    207.029k (± 2.5%) i/s -      1.043M in   5.039794s
            #members      4.615M (± 0.7%) i/s -     23.111M in   5.008104s
```

After
```
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [arm64-darwin21]
Warming up --------------------------------------
             .define    21.214k i/100ms
            #members     1.818M i/100ms
Calculating -------------------------------------
             .define    210.958k (± 2.3%) i/s -      1.061M in   5.030760s
            #members     18.104M (± 0.2%) i/s -     90.911M in   5.021470s
```